+++ engine/include/analyzer/report/report.h
@@ -10,6 +10,7 @@
 #include "engine/include/analyzer/report/instr.h"
 #include "engine/include/analyzer/report/sfr.h"
 #include "engine/include/analyzer/report/label.h"
+#include "engine/include/analyzer/report/vector.h"
 
 typedef struct vmcu_report {
     
@@ -22,9 +23,12 @@ typedef struct vmcu_report {
     int32_t n_label;
     vmcu_label_t *label;
 
+    int32_t n_vector;
+    vmcu_vector_t *vector;
+
 } vmcu_report_t;
 
-extern vmcu_report_t* vmcu_report_ctor(const char *file);
+extern vmcu_report_t* vmcu_report_ctor(void);
 extern void vmcu_report_dtor(vmcu_report_t *this);
 
 #endif