+++ engine/include/analyzer/util/util.h
@@ -1,0 +1,10 @@
+/* Analyzer Utilities Header */
+
+#ifndef VMCU_UTIL_H
+#define VMCU_UTIL_H
+
+typedef struct vmcu_instr vmcu_instr_t;
+
+extern int vmcu_get_abs_addr(const vmcu_instr_t *instr);
+
+#endif