+++ engine/src/analyzer/analyzer.c
@@ -7,6 +7,7 @@
 #include "engine/include/analyzer/analyzer.h"
 #include "engine/include/analyzer/modules/sfr/sfr_analyzer.h"
 #include "engine/include/analyzer/modules/labels/label_analyzer.h"
+#include "engine/include/analyzer/modules/vector/vector_analyzer.h"
 #include "engine/include/analyzer/report/report.h"
 #include "engine/include/disassembler/disassembler.h"
 
@@ -17,7 +18,7 @@
 
 vmcu_report_t* vmcu_analyze_ihex(const char *hex_file, vmcu_model_t *mcu) {
 
-    vmcu_report_t *report = vmcu_report_ctor(hex_file);
+    vmcu_report_t *report = vmcu_report_ctor();
     report->disassembly = vmcu_disassemble_ihex(hex_file, &report->progsize, mcu);
 
     if(report->disassembly == NULL) {
@@ -29,6 +30,7 @@ vmcu_report_t* vmcu_analyze_ihex(const char *hex_file, vmcu_model_t *mcu) {
     /* --- architecture specific analysis --- */
 
     vmcu_analyze_sfr(report, mcu);
+    vmcu_analyze_vectors(report, mcu);
 
     /* --- architecture unspecific analysis --- */